+++ Client/src/ClientBody.h
@@ -1,5 +1,4 @@
 #pragma once
-#include <cstdint>
 
 namespace LSMS {
