+++ common/CommonInclude.h
@@ -1,0 +1,9 @@
+#pragma once
+
+//LibMsg Include
+#include <vector>
+#include <cstdint>
+#include <LibMsg.h>
+
+//Communication
+#include <asio.hpp>
