+++ common/Time.cpp
@@ -1,0 +1,12 @@
+#include "Time.h"
+
+namespace LSMS
+{
+    char *libtime::getdate_mmddyyyy(char *_str)
+    {
+        std::time_t t = std::time(0);
+        strftime(_str, MAXDATELENGTH, "%m%d%Y", localtime(&t));
+        return _str;
+    }
+
+}
