+++ common/Time.h
@@ -1,0 +1,14 @@
+#pragma once
+#include <ctime>
+
+namespace LSMS
+{
+    namespace libtime
+    {
+        constexpr size_t MAXDATELENGTH = 80;
+
+        char *getdate_mmddyyyy(char *);
+
+    }
+
+};
