--- src/Time.h
@@ -1,10 +1,0 @@
-#pragma once
-#include <ctime>
-
-namespace libtime
-{
-    constexpr size_t MAXDATELENGTH = 80;
-
-    char *getdate_mmddyyyy(char *);
-
-}
